requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
attrs==25.3.0
audioop-lts==0.2.1
binaryornot==0.4.4
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
cryptography==45.0.5
discord.py==2.5.2
dotenv==0.9.9
fastapi==0.116.1
Flask==3.1.1
flask-cors==6.0.1
flet==0.28.3
flet-cli==0.28.3
flet-desktop==0.28.3
flet-web==0.28.3
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
Jinja2==3.1.6
keyring==25.6.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
multidict==6.5.0
oauthlib==3.3.1
packaging==25.0
propcache==0.3.2
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
PyGithub==2.7.0
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pypng==0.20220715.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
requests==2.32.4
rich==14.0.0
Send2Trash==1.8.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.2
text-unidecode==1.3
toml==0.10.2
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchdog==4.0.2
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
